add security setting and minor change
