+++ selectfood/src/main/java/com/project/selectfood/config/SelectfoodConfiguration.java
@@ -11,7 +11,7 @@ import org.springframework.context.annotation.Configuration;
 public class SelectfoodConfiguration {
 
     @Bean
-    public ObjectMapper objectMapper(){
+    public ObjectMapper objectMapper() {
         ObjectMapper mapper = new ObjectMapper();
         mapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
         mapper.setVisibility(VisibilityChecker.Std.defaultInstance().withFieldVisibility(JsonAutoDetect.Visibility.ANY));
