--- selectfood/src/main/java/com/project/selectfood/config/SelectfoodSecurityConfiguration.java
@@ -1,21 +1,0 @@
-package com.project.selectfood.config;
-
-import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
-@EnableWebSecurity
-public class SelectfoodSecurityConfiguration extends WebSecurityConfigurerAdapter {
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http
-                .authorizeRequests()
-                .antMatchers(HttpMethod.GET, "/selectfood/**").permitAll()
-                .antMatchers(HttpMethod.POST, "/selectfood/users").authenticated()
-                .anyRequest().authenticated()
-                .and()
-                .csrf().disable();
-    }
-}
