+++ selectfood/src/main/java/com/project/selectfood/models/AdditionalItem.java
@@ -1,5 +1,4 @@
-package com.project.selectfood.data;
-
+package com.project.selectfood.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,21 +6,21 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotEmpty;
-import java.util.Date;
 
 @Entity
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class FindingHistory {
+public class AdditionalItem {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long code;
 
+    @Column(unique = true)
     @NotEmpty(message = "Restaurant name")
     private String name;
 
-    @Column(unique = true)
-    private Date date;
+    @Column(columnDefinition = "int default 1")
+    private int count;
 }
