+++ selectfood/src/main/java/com/project/selectfood/models/ERole.java
@@ -1,7 +1,7 @@
 package com.project.selectfood.models;
 
 public enum ERole {
-  ROLE_USER,
-  ROLE_MODERATOR,
-  ROLE_ADMIN
+    ROLE_USER,
+    ROLE_MODERATOR,
+    ROLE_ADMIN;
 }
