+++ selectfood/src/main/java/com/project/selectfood/models/Role.java
@@ -19,5 +19,4 @@ public class Role {
   @Enumerated(EnumType.STRING)
   @Column(length = 20)
   private ERole name;
-
 }
