+++ selectfood/src/main/java/com/project/selectfood/models/User.java
@@ -9,6 +9,7 @@ import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 @Entity
@@ -43,6 +44,12 @@ public class User {
             joinColumns = @JoinColumn(name = "user_id"),
             inverseJoinColumns = @JoinColumn(name = "role_id"))
     private Set<Role> roles = new HashSet<>();
+
+    @OneToMany
+    private List<AdditionalItem> additionalItems;
+
+    @OneToMany
+    private List<FindingHistory> findingHistories;
 
     public User(String username, String email, String password) {
         this.username = username;
