+++ selectfood/src/main/java/com/project/selectfood/repository/AdditionItemsRepo.java
@@ -1,6 +1,6 @@
 package com.project.selectfood.repository;
 
-import com.project.selectfood.data.AdditionalItem;
+import com.project.selectfood.models.AdditionalItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdditionItemsRepo extends JpaRepository<AdditionalItem, Long> {
