+++ selectfood/src/main/java/com/project/selectfood/repository/FindingHistoryRepo.java
@@ -1,6 +1,6 @@
 package com.project.selectfood.repository;
 
-import com.project.selectfood.data.FindingHistory;
+import com.project.selectfood.models.FindingHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FindingHistoryRepo extends JpaRepository<FindingHistory, Long> {
