+++ selectfood/src/main/resources/data.sql
@@ -1,0 +1,3 @@
+INSERT INTO ROLES (id, name) VALUES (1, 'ROLE_USER');
+INSERT INTO ROLES (id, name) VALUES (2, 'ROLE_MODERATOR');
+INSERT INTO ROLES (id, name) VALUES (3, 'ROLE_ADMIN');
